requirements.txt
# Core dependencies
python-dotenv>=1.0.0
chromadb>=0.4.0
rank-bm25>=0.2.2
gitpython>=3.1.0

# Tree-sitter for AST parsing
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-javascript>=0.20.0
tree-sitter-typescript>=0.20.0
tree-sitter-go>=0.20.0
tree-sitter-java>=0.20.0

# LLM providers (install what you need)
anthropic>=0.8.0  # For Claude
openai>=1.0.0     # For OpenAI
google-generativeai>=0.3.0  # For Gemini

# Neo4j (optional but recommended)
neo4j>=5.0.0

# Utilities
networkx>=3.0  # Fallback graph if Neo4j unavailable
requests>=2.31.0
tqdm>=4.65.0
